Part 5: refactor
